backend/requirements.txt
gunicorn==21.2.0
python-dotenv==1.0.1
moviepy==1.0.3
pydub==0.25.1 
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
cachelib==0.13.0
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
distro==1.9.0
exceptiongroup==1.2.2
filelock==3.18.0
Flask==3.0.2
Flask-Cors==4.0.0
Flask-Session==0.5.0
Flask-SQLAlchemy==3.1.1
fsspec==2025.3.2
greenlet==3.2.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.7.0
mpmath==1.3.0
networkx==3.4.2
numba==0.61.2
numpy==2.2.5
openai==1.76.2
openai-whisper==20240930
pillow==11.0.0
psutil==7.0.0
pydantic==2.11.4
pydantic_core==2.33.2
regex==2024.11.6
requests==2.32.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
sympy==1.14.0
tiktoken==0.9.0
torch==2.7.0
torchaudio==2.7.0+cpu
torchvision==0.22.0+cpu
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
